slow it down take a breath it's alright

+++ helpers.c
@@ -1,6 +1,6 @@
 #include "codexion.h"
 
-long get_time_ms(void)
+long long get_time_ms(void)
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
